refactor param bindings to put extra functions into source files
