+++ library/python/CMakeLists.txt
@@ -21,9 +21,14 @@ if( Scarab_BUILD_PYTHON )
 
     if( Scarab_BUILD_PARAM )
         add_definitions( -DBUILD_PARAM_PYBINDING )
+        set( PYBINDING_SOURCEFILES
+             ${PYBINDING_SOURCEFILES}
+             ${dir}/param_binding_helpers.cc
+           )
         set( PYBINDING_HEADERFILES
             ${PYBINDING_HEADERFILES}
             ${dir}/param_pybind.hh
+            ${dir}/param_binding_helpers.hh
             ${dir}/param_value_pybind.hh
             ${dir}/param_array_pybind.hh
             ${dir}/param_node_pybind.hh
